requirements.txt
# Core Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
gitpython>=3.1.0
pytest>=7.0.0
typing-extensions>=4.8.0
pydantic>=2.0.0
pydantic-settings>=2.1.0
requests>=2.31.0
google-generativeai>=0.8.0

# Caching
diskcache>=5.6.0

# Logging & Observability
structlog>=23.2.0
prometheus-client>=0.19.0
rich>=13.7.0

# Rate Limiting & Resilience
tenacity>=8.2.0
circuitbreaker>=1.4.0
